+++ weaver/utils.py
@@ -28,7 +28,6 @@ from pyramid.exceptions import ConfigurationError
 from pyramid.httpexceptions import HTTPError as PyramidHTTPError, HTTPGatewayTimeout, HTTPTooManyRequests
 from pyramid.registry import Registry
 from pyramid.request import Request as PyramidRequest
-from pyramid.response import _guess_type as guess_file_contents  # noqa: W0212
 from pyramid.settings import asbool, aslist
 from pyramid.threadlocal import get_current_registry
 from pyramid_beaker import set_cache_regions_from_settings
@@ -377,52 +376,6 @@ def localize_datetime(dt, tz_name="UTC"):
     else:
         warnings.warn("tzinfo already set", TimeZoneInfoAlreadySetWarning)
     return tz_aware_dt
-
-
-def get_file_header_datetime(dt):
-    # type: (datetime) -> str
-    """
-    Obtains the standard header datetime representation.
-
-    .. seealso::
-        Format of the date defined in :rfc:`5322#section-3.3`.
-    """
-    dt_gmt = localize_datetime(dt, "GMT")
-    dt_str = dt_gmt.strftime("%a, %d %b %Y %H:%M:%S GMT")
-    return dt_str
-
-
-def get_file_headers(path, download_headers=False, content_headers=False, content_type=None):
-    # type: (str, bool, bool, Optional[str]) -> HeadersType
-    """
-    Obtain headers applicable for the provided file.
-
-    :param path: File to describe.
-    :param download_headers: If enabled, add the attachment filename for downloading the file.
-    :param content_headers: If enabled, add ``Content-`` prefixed headers.
-    :param content_type: Explicit ``Content-Type`` to provide. Otherwise, use default guessed by file system.
-    :return: Headers for the file.
-    """
-    stat = os.stat(path)
-    headers = {}
-    if content_headers:
-        c_type, c_enc = guess_file_contents(path)
-        headers.update({
-            "Content-Type": content_type or c_type,
-            "Content-Encoding": c_enc or "",
-            "Content-Length": str(stat.st_size)
-        })
-        if download_headers:
-            headers.update({
-                "Content-Disposition": f"attachment; filename=\"{os.path.basename(path)}\"",
-            })
-    f_modified = get_file_header_datetime(datetime.fromtimestamp(stat.st_mtime))
-    f_created = get_file_header_datetime(datetime.fromtimestamp(stat.st_ctime))
-    headers.update({
-        "Date": f_created,
-        "Last-Modified": f_modified
-    })
-    return headers
 
 
 def get_base_url(url):